+++ sugarloaf/src/font/constants.rs
@@ -15,7 +15,7 @@ pub const DEFAULT_FONT_FAMILY: &str = "cascadiacode";
 // CascadiaCode-Italic.ttf
 // CascadiaCode-Light.ttf
 // CascadiaCode-LightItalic.ttf
-// CascadiaCode-Regular.ttf
+// CascadiaCodeNF-Regular.ttf
 // CascadiaCode-SemiBold.ttf
 // CascadiaCode-SemiBoldItalic.ttf
 // CascadiaCode-SemiLight.ttf
@@ -43,7 +43,7 @@ pub const FONT_CASCADIAMONO_LIGHT_ITALIC: &[u8] =
     font!("./resources/CascadiaCode/CascadiaCode-LightItalic.otf");
 
 pub const FONT_CASCADIAMONO_REGULAR: &[u8] =
-    font!("./resources/CascadiaCode/CascadiaCode-Regular.otf");
+    font!("./resources/CascadiaCode/CascadiaCodeNF-Regular.otf");
 
 pub const FONT_CASCADIAMONO_SEMI_BOLD: &[u8] =
     font!("./resources/CascadiaCode/CascadiaCode-SemiBold.otf");
@@ -58,6 +58,6 @@ pub const FONT_CASCADIAMONO_SEMI_LIGHT_ITALIC: &[u8] =
     font!("./resources/CascadiaCode/CascadiaCode-SemiLightItalic.otf");
 
 pub const FONT_SYMBOLS_NERD_FONT_MONO: &[u8] =
-    font!("./resources/SymbolsNerdFontMono/SymbolsNerdFontMono-Regular.ttf");
+    font!("./resources/CascadiaCode/CascadiaCodeNF-Regular.otf");
 
 pub const FONT_TWEMOJI_EMOJI: &[u8] = font!("./resources/Twemoji/Twemoji.Mozilla.ttf");