+++ docs/docs/releases.md
@@ -154,6 +154,9 @@ These changes are particularly beneficial for:
 
 - Fix: In some cases, the first typed character doesn't display until after a delay, or until another key is hit [#1098](https://github.com/raphamorim/rio/issues/1098).
 - Fix: Anomalous behavior occurs with the Bookmark tab style in the new versions 0.14 and 0.13. [#1094](https://github.com/raphamorim/rio/issues/1094).
+- Support private user area (NF) auto decrease of width in case of next cell is occupied.
+- Replace Cascadia Code Regular with Cascadia Code NF Regular.
+- Remove Symbols Nerd Font Mono font in favor of Cascadia Code NF Regular.
 
 ## 0.2.14